Refactor with a components folder
