+++ src/App.css
@@ -40,3 +40,8 @@
 .read-the-docs {
   color: #888;
 } */
+
+img {
+  width: 100%;
+  max-width: 100%;
+}
