+++ src/components/layout/hero/Hero.css
@@ -1,0 +1,10 @@
+.hero {
+  background: #d40000;
+  width: 100vw;
+  max-width: 100vw;
+  max-height: 40vh;
+}
+
+.hero img {
+  max-width: 50%;
+}
