+++ src/components/layout/hero/Hero.jsx
@@ -1,0 +1,15 @@
+import "./Hero.css";
+import HeroImage from "/images/Hero-Image-Mobile.jpg";
+
+function Hero() {
+  return (
+    <section className="hero">
+      <p>Hey, my name is</p>
+      <h1>Caleb Anderson</h1>
+      <p>Full Stack Developer & Engineer</p>
+      <img src={HeroImage} alt="Hero image placeholder" />
+    </section>
+  );
+}
+
+export default Hero;
